+++ tests/test_playwright.py
@@ -4,7 +4,6 @@ import unittest
 
 from app.common import TIMESTAMP_LONG
 from app.pagedriver import get_browser_context
-from app.scraper.info.igdb import get_igdb_details, get_igdb_id
 from app.scraper.info.steam import get_steam_details, get_steam_id
 
 logging.basicConfig(
@@ -20,39 +19,6 @@ class PlaywrightTests(unittest.IsolatedAsyncioTestCase):
             page = await context.new_page()
             res = await page.goto("https://google.com/")
             self.assertEqual(res.status, 200)
-
-
-class IGDBGameInfoTests(unittest.IsolatedAsyncioTestCase):
-    async def test_igdb_id_resolution(self) -> None:
-        expected_id: int = 7360  # Tom Clancy's Rainbow Six® Siege
-        scraped_id: int = await get_igdb_id(
-            "Rainbow Six Siege",
-        )
-        self.assertEqual(expected_id, scraped_id)
-
-    async def test_igdb_id_resolution_with_special_chars(self) -> None:
-        expected_id: int = 66
-        scraped_id: int = await get_igdb_id(
-            "Monkey Island 2 Special Edition: LeChuck’s Revenge",
-        )
-        self.assertEqual(expected_id, scraped_id)
-
-    async def test_igdb_details_counterstrike(self) -> None:
-        igdb_info = await get_igdb_details(
-            title="Counter-Strike",
-        )
-        self.assertIsNotNone(igdb_info)
-        self.assertEqual(igdb_info.name, "Counter-Strike")
-        self.assertIsNotNone(igdb_info.short_description)
-        self.assertEqual(
-            igdb_info.release_date.isoformat(), "2000-11-09T00:00:00+00:00"
-        )
-        self.assertGreater(igdb_info.meta_ratings, 1)
-        self.assertGreater(igdb_info.meta_score, 50)
-
-        self.assertGreater(igdb_info.user_ratings, 400)
-        self.assertGreater(igdb_info.user_score, 50)
-        self.assertEqual(igdb_info.url, "https://www.igdb.com/games/counter-strike")
 
 
 class SteamGameInfoTests(unittest.IsolatedAsyncioTestCase):
